requirements.txt
google-genai
python-dotenv
python-multipart
fastapi
uvicorn